+++ src/controllers/ScheduleTypeController.ts
@@ -3,7 +3,16 @@ import ScheduleType from "../models/ScheduleType";
 
 export const getScheduleTypeById = async (req: Request, res: Response) => {
     try {
-        const scheduleType = await ScheduleType.findByPk(req.params.id);
+        const { id } = req.params;
+
+        // Validar entrada
+        if (!id) {
+            return res
+                .status(400)
+                .json({ message: "ScheduleType ID is required" });
+        }
+
+        const scheduleType = await ScheduleType.findByPk(id);
         if (scheduleType) {
             res.status(200).json(scheduleType);
         } else {
@@ -40,13 +49,21 @@ export const createScheduleType = async (req: Request, res: Response) => {
 
 export const updateScheduleType = async (req: Request, res: Response) => {
     try {
-        const [updated] = await ScheduleType.update(req.body, {
-            where: { UniqueID: req.params.id },
+        const { id } = req.params;
+
+        // Validar entrada
+        if (!id) {
+            return res.status(400).json({ message: "Event ID is required" });
+        }
+
+        // Excluir campos no permitidos para actualización
+        const { id: bodyId, ...updateFields } = req.body;
+
+        const [updated] = await ScheduleType.update(updateFields, {
+            where: { id },
         });
         if (updated) {
-            const updatedScheduleType = await ScheduleType.findByPk(
-                req.params.id
-            );
+            const updatedScheduleType = await ScheduleType.findByPk(id);
             res.status(200).json(updatedScheduleType);
         } else {
             res.status(404).json({ message: "ScheduleType not found" });
@@ -58,8 +75,17 @@ export const updateScheduleType = async (req: Request, res: Response) => {
 
 export const deleteScheduleType = async (req: Request, res: Response) => {
     try {
+        const { id } = req.params;
+
+        // Validar entrada
+        if (!id) {
+            return res
+                .status(400)
+                .json({ message: "ScheduleType ID is required" });
+        }
+
         const deleted = await ScheduleType.destroy({
-            where: { UniqueID: req.params.id },
+            where: { id },
         });
         if (deleted) {
             res.status(204).json({ message: "ScheduleType deleted" });
